rules_python_setup: default use_pip to false
It seems more conservative to opt out of a ruleset by default unless specifically requested. Users of Python rules who don't need packaging support won't necessarily know to disable this flag for a more minimal build. In any case, we'll make the usage clear in the rules_python readme once the federation has settled.

We could flip this back to true at some point in the future if we believe the packaging rules are stable.