rules_python: Move non-distro targets to a separate presubmit config
As of #75 the federation references a rules_python release instead of the entire repository. However, the presubmit failed (as expected) since some test targets cover files that are not part of the distribution.
The present commit moves these tests to a separate file so that they can be tested again once #78 is solved.

Tested: https://buildkite.com/bazel/fwes-federation-pipeline/builds/10