Conditionally add internal deps and setup to WORKSPACE.
Whenever we test a release of a member project there is no need to call the internal deps and setup functions.
Consequently, this commit allows to control the behavior with the --internal flag.
Ultimately the flag should default to False, but for now we use True to guarantee backwards compatibility.
This commit uses integers instead of boolean values due to problems with arg_parse.

Required by #78